Exported model and cleaned up repo

+++ ML Models/configs.py
@@ -1,5 +1,5 @@
 import torch
 
-datatype = torch.float64
+datatype = torch.float32
 #datatype = torch.complex128
 window_size = 128
